1882. Process Tasks Using Servers
